@@ -25,8 +25,8 @@
  */
 package chat.dim.ui.grid;
 
-import android.support.annotation.NonNull;
-import android.support.v4.os.TraceCompat;
+import androidx.annotation.NonNull;
+import androidx.core.os.TraceCompat;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.BaseAdapter;
